Consolidate suite runs for docker
