+++ internal/docker/container.go
@@ -2,6 +2,8 @@ package docker
 
 import (
 	"context"
+	"fmt"
+	"github.com/rs/zerolog/log"
 	"github.com/saucelabs/saucectl/cli/command"
 )
 
@@ -12,4 +14,28 @@ type ContainerRunner struct {
 	containerID     string
 	docker          *Handler
 	containerConfig *containerConfig
+}
+
+func (r *ContainerRunner) run(cmd []string, env map[string]string) error {
+	defer func() {
+		log.Info().Msg("Tearing down environment")
+		if err := r.docker.Teardown(r.Ctx, r.containerID); err != nil {
+			if !r.docker.IsErrNotFound(err) {
+				log.Error().Err(err).Msg("Failed to tear down environment")
+			}
+		}
+	}()
+
+	exitCode, err := r.docker.ExecuteAttach(r.Ctx, r.containerID, r.Cli, cmd, env)
+	log.Info().
+		Int("ExitCode", exitCode).
+		Msg("Command Finished")
+
+	if err != nil {
+		return err
+	}
+	if exitCode != 0 {
+		return fmt.Errorf("exitCode is %d", exitCode)
+	}
+	return nil
 }
